// gang/Week10/Week10/src/types/movies.ts
export type requestSearchMovie ={ 
    query:string,
    include_adult:boolean,
    language:string
}

export type responseMovie = {
    page: number,
    results:Movie[],
    total_pages:number,
    total_results:number, 
}

export type Movie = {
  adult: boolean;
  backdrop_path: string;
  genre_ids: number[];
  id: number;
  original_language: string;
  original_title: string;
  overview: string;
  popularity: number;
  poster_path: string;
  release_date: string;
  title: string;
  video: boolean;
  vote_average: number;
  vote_count: number;
};